Create Poor pigs
link: https://leetcode.com/problems/poor-pigs/

With 2 pigs, poison killing in 15 minutes, and having 60 minutes, we can find the poison in up to 25 buckets in the following way. Arrange the buckets in a 5×5 square:

 1  2  3  4  5
 6  7  8  9 10
11 12 13 14 15
16 17 18 19 20
21 22 23 24 25

the first pig will drink the first row and the second will drink the first column then the worst case both do not die and goes for the second row and second column. Similarly, if we have 3 pigs then in this case only 5*5*5 pigs can get tested within time.

credit: @StefanPochmann