Create Bouncy Ball
Here normal DFS can help.
x->row
y->col
if the given direction's first character is 'U' it means we are going in the up direction. so x will decrease for sure [dx=-1] else [dx=1]
if the second character is 'R' it means that we are going in the right direction. so y will decrease[dy=-1] else [dy=1]
Then check if we are at the first row  and dx=-1 which will be a bounce  or if we are at the last row and dx =1 which will be a bounce we need to change x directions so make dx as  -dx
Again check if we are at the first column and dy = -1 which will be a bounce or if we are at the last column and dy =1 which will also be a bounce, we need to change y direction, i.e, dy as -dy
Hence continue with the current position and the modified direction.
If we got a position that has already been visited with the current direction then return -1 and made flag = false that no path is there to reach the destination.